cojp_cbor: Rename cbor.[ch] tp cojp_cbor.[ch]
