+++ openapps/SConscript
@@ -33,7 +33,7 @@ additionalSourceFiles = [
     os.path.join('opencoap','hmac.c'),
     os.path.join('opencoap','sha224-256.c'),
     os.path.join('opencoap','cborencoder.c'),
-    os.path.join('cjoin','cbor.c'),
+    os.path.join('cjoin','cojp_cbor.c'),
 ]
 
 # header files that should be included in the build other than
@@ -42,7 +42,7 @@ additionalHeaderFiles = [
     os.path.join('opencoap','openoscoap.h'),
     os.path.join('opencoap','sha.h'),
     os.path.join('opencoap','cborencoder.h'),
-    os.path.join('cjoin','cbor.h'),
+    os.path.join('cjoin','cojp_cbor.h'),
 ]
 
 # additional apps the user wants to build
